+++ RELEASE_NOTES.md
@@ -1,8 +1,8 @@
 # Release Notes
 
-## 0.9.0 Series
+## 0.10.0 Series
 
-### 0.9.1 - XX-XX-20XX
+### 0.10.0 - XX-XX-20XX
 
 * Fixed performance issue related to Nagle's algorithm on RTEMS 4.X's legacy networking stack
     * `TCP_NODELAY` is now toggled when needed to improve write and read performance.
@@ -11,6 +11,11 @@
 * Improved logging
 * Avoid allocating POSIX condition variables and mutexes for transactions when they're not needed.
 * Fixed out-of-bounds read in `p9Mount` when the `uid.gid` portion of the IP string is not provided.
+* Removed CMake build support
+* Added waf build support with integration for RTEMS 4
+* Fixed typo in verbose logging
+
+## 0.9.0 Series
 
 ### 0.9.0 - 09-30-2025